Made libmatchbox2 build on non-32-bit systems

+++ matchbox/client-types/mb-wm-client-input.c
@@ -12,7 +12,7 @@ mb_wm_client_input_request_geometry (MBWindowManagerClient *client,
 				     MBWMClientReqGeomType  flags);
 
 static void
-mb_wm_client_input_stack (MBWindowManagerClient *client, int flags);
+mb_wm_client_input_stack (MBWindowManagerClient *client, intptr_t flags);
 
 static void
 mb_wm_client_input_class_init (MBWMObjectClass *klass)
@@ -140,7 +140,7 @@ mb_wm_client_input_request_geometry (MBWindowManagerClient *client,
 }
 
 static void
-mb_wm_client_input_stack (MBWindowManagerClient *client, int flags)
+mb_wm_client_input_stack (MBWindowManagerClient *client, intptr_t flags)
 {
   MBWM_MARK();
